+++ nullaway/src/test/java/com/uber/nullaway/NullAwayTest.java
@@ -3227,25 +3227,4 @@ public class NullAwayTest {
             "}")
         .doTest();
   }
-
-  @Test
-  public void mapReassignUnsound() {
-    defaultCompilationHelper
-        .addSourceLines(
-            "Test.java",
-            "package com.uber;",
-            "import java.util.*;",
-            "public class Test {",
-            "  public void mapReassignUnsound(Map m, Object o) {",
-            "    if (m.containsKey(o)) {",
-            "      // NullAway is currently unsound for this case.  It ignores",
-            "      // the re-assignment of m and still assumes m.get(o) is non-null",
-            "      // on the subsequent line.",
-            "      m = new HashMap();",
-            "      m.get(o).toString();",
-            "    }",
-            "  }",
-            "}")
-        .doTest();
-  }
 }
